+++ index.html
@@ -229,7 +229,7 @@
       <button type="button" class="error__button">Попробовать снова</button>
     </div>
   </template>
-  <script src='main.js'></script>
+  <script src='main.js' type='module'></script>
 
 </body>
 </html>
